Implementing LOG4J2-155: add getFormat to Layout
https://issues.apache.org/jira/browse/LOG4J2-155

Exposing an advertiser attribute on existing configuration implementations and providing a default Multicast DNS advertiser implementation.

Implementing support for advertisement in socket-based appenders, FileAppender and RollingFileAppender.

While the Advertiser implementation class is specified on the configuration, enabling advertisement is on a per-appender basis, and may require additional attributes (for example, FileAppender requires and advertiseURI attribute).

With a follow-on update to Chainsaw, Chainsaw will be able to discover the published appender entries and automatically begin tailing the appender with no user configuration required.

Also committed tests which verified:
  - only appenders with 'advertise=true' are actually advertised
  - the appender is 'unadvertised' when configuration.stop() is called
  - the appender is re-advertised when configuration.reconfigure() is subsequently called
  
  Updated docs in the 'configuration' section of the manual

git-svn-id: https://svn.apache.org/repos/asf/logging/log4j/log4j2/trunk@1451133 13f79535-47bb-0310-9956-ffa450edef68