Expose the XML or JSON configuration text via the Advertiser mechanism.
MulticastDNSAdvertiser has a limit of 255 bytes for labels and values.  If an entry is greater than 255 bytes, the entry is now removed prior to publishing.

The log4j configuration content was exceeding this limit.  This limit is specific to Multicast DNS.

Added a 'location' key, with the location (file path) if it is available


git-svn-id: https://svn.apache.org/repos/asf/logging/log4j/log4j2/trunk@1480885 13f79535-47bb-0310-9956-ffa450edef68