LOG4J2-673 – plugin preloading fails in shaded jar files Support for plugin preloading through the standard javax.annotation.processing tool was adding in LOG4J2-595
But the plugin processor always creates and stores the processed "Plugin" annotated classes into the same file. This works fine when the classpath consists of individual jar files, but fails when shaded jar files are used.

The fix saves the dat file in a location under META-INF that matches the shared package all the processed plugins are found under.
The package attribute in the config file is then used so that multiple dat files can be loaded at runtime.
This means that the package attribute is no longer deprecated.