LOG4J2-493 Wrap-up: better web app support for async loggers.
It is now possible to place the logging jars in the container's
classpath when making all loggers asynchronous by using
AsyncLoggerContextSelector. Web apps can be started and stopped.

- handle the case where the log4j jars are in container's classpath, but
the log4j2.xml config file is in a web app's classpath. In this
scenario, the Disruptor should only be started that web app.