LOG4J2-435 attempt to fix failing test on unix
Sorting seems to be broken. I suspect this is causing the test failures.
Try taking a snapshot of BasicFileAttributes values during initial file
tree walk instead of just keeping a reference to the attributes object
itself: this object may just be a view whose values change during the
tree walk.