LOG4J2-1300 - Remove serializability from classes that don't need it
Appenders and Filters are already supposed to be Serializable based on the abstract classes. This is not feasible based on the use of many non-serializable JDK classes (e.g., OutputStream, Buffer). As LogEvent, Logger, and MessageFactory (implementations) all implement Serializable without requiring everything else, then there's no need to serialize appenders, filters, and all the other plugin objects that are never serialized anywhere.