Documentation bug for purgeAscending method.
For "@param highIndex  Log file associated with high index will be deleted if needed. "
That is incorrect it seems, that applies to purgeDescending.

That message should be under "@param lowIndex with the message Log file associated with low index will be deleted if needed. "

Because purgeAscending deletes the one with the low index if required, if I'm not wrong.