[LOG4J2-1730]: Add literal values and date type support to CassandraAppender
The previously assumed method of encoding dates as a long value does not work properly. Instead, this commit adds the ability to use a java.util.Date-compatible type conversion for ColumnMapping which uses the LogEvent timestamp instead of a StringLayout->TypeConverter chain.

This also adds the ability to use literal values in column mappings similar to the equivalent feature in JdbcAppender.