LOG4J2-2527: ListAppender getters return immutable snapshots
Provides a bit more thread safety to ListAppender. The class
is marked not-thread-safe, however this functionality is
important for testing, particularly when asynchronous logging is
configured.

Previously we returned unmodifiable wrappers around the internal
state, which can be dangerous to iterate over while events are
actively logged.

State is tracked in synchronized lists, which should be safer
than synchronizing on individual methods.