LOG4J-2735 - make PluginCache output reproducible
Previously, the data written by a PluginCache with identical
categories and entries would vary depending on the order in which
those categories or entries were added to the cache. The ordering of
the elements handled by the annotation processor that populates the
cache is not guaranteed so multiple compilations of identical source
code could result in different output. This prevented projects that
define Log4j2 plugins from having a reproducible build.

This commit updates PluginCache to use TreeMaps to hold categories
and their entries. This ensures that the cache's data is stored and
written in an order determined by sorting their String keys rather
than by when they are added. This makes the cache's output
reproducible, irrespective of the order of the elements handled by
the annotation processor.