LOG4J2-2768 Add log(String, ...) overloads to LogBuilder
The overloaded methods behave similarly to log(String, Object[]), but reduce the need for an Object[] allocation in certain code paths. The added methods and their Javadoc come from the Logger class. The Javadoc have been modified with the additional of the @see reference to the Unbox class. The Unbox class did not exist when the Logger Javadoc were originally written, and it seems pertinent to make users aware of the performance benefit when using these methods.