Conditionally perform status logging calculations in PluginRegistry
Debug status logging is typically disabled, but formatting calculations were performed regardless.

Moving the formatting logic inside of a conditional reduces the startup time of applications. The reduction is mostly attributable to the fact that the class initialization of DecimalFormat is non-trivial for the JVM, as it loads a variety of formatting information.