Add logic to detect kafka broker is valid or not
Add logic to detect kafka broker is valid or not during app startup.
User should be noticed with kafka broker available or not.
Mark kafka appenders as not disabled if kafka broker not available. Then Logevents won't be stored in this appender.
Otherwise, the kafka producer will retry unlimited with lots of log. If user didn't add blocking parameter to asyncappender, all appenders in this logger will be affected.