JTL memory optimization: avoid additional string builder
JsonWriter avoids an additional StringBuilder for formatting in
favor of escaping in places.
It's worth some benchmarks to compare and contrast. In the ideal
case, strings don't contain characters that must be escaped, and no
memory copying is necessary. When escaping does occur, contents are
localized to an array and cache hits are likely. We may not need
to reach into main memory as frequently, unless we've escaped enough
characters to fill a cache line.

This commit is a proof of concept, it's not ideal that we're combining
two separate json string escaping implementations in JsonWriter:
`StringBuilders.escapeJson` and `JsonWriter.quoteString`