LOG4J2-3103: Make listeners in LoggerContext a CopyOnWriteArrayList (#508)
This particular case fails quite often in integration tests
where you have multiple separate threads potentially creating
and destroying the singleton LoggerContext independently.

Since this is a Listener List, it makes sense to just make
it a CopyOnWriteArrayList, since the frequency of both
occurring is going to be roughly equivalent.

Co-authored-by: Mike Glazer <mglazer@palantir.com>