Prefer `indexOf(char)`
Note that indexOf(char) performs worse than indexOf(String) on pre-16
JREs due to missing intrinsics for the character implementation.
The difference is a few nanoseconds in most cases, so we opt to give the
jre as much information as possible for best performance on new
runtimes, with the possibility that such optimizations may be
back-ported.

See https://bugs.openjdk.java.net/browse/JDK-817358