[DOC] fix Duration string
According to
https://logging.apache.org/log4j/2.x/log4j-core/apidocs/org/apache/logging/log4j/core/appender/rolling/action/Duration.html#parseCharSequence
a Duration looks like "P2D" or "PT20S".

In serveral instances our configuration did not work as expected and
didn't delete files despite them beeing older then the period that we
thought was configured. After some digging into the configuration we
found that the above mention string fixed the configuration.