[LOG4J2-3326] Fix parsing filters from Log4j 1.2 properties configuration file (#680)
* Use specific filter key as prefix while parsing configuration

Problem: The log4j 1.x configuration file is not properly parsed, in regards of filters. Whatever is configured in file, is getting ignored, and default values are used. 
E.g. for "level range filter" the LevelMax is not properly found (eventhough properly read as property).

Root cause: The default filter prefix is used, instead of specific filter key.

* add test, verifying LevelRangeFilter properly parsed/configured

...with values of properties configuration file