Expand test suite and fix NPE
This commit strengthens the test suite for the `StructuredDataLookup` class so that it achieves full branch coverage.

The corner cases revealed the problem that a lookup of a null key with a correct event led to a null pointer exception. This is inconsistent with the javadoc ("can be null") and the other lookup classes.

The commit fixes the null pointer exception for the `key` by replacing `key.equals("id")` by `"id".equals(key)`, in two places.

It also introduces explicit constants for the string literals "id" and "type".