LOG4J2-3317: Fix RoutingAppender backcompat and improve lookup security
This change unifies our configuration-time and runtime property
handling logic in order to make configuration-time properties
safer and more obvious, while allowing RoutingAppender configurations
that were broken by 2.17.1 to work as expected once again.

This is done using two high-level ideas:
1. Properties defined in the configuration are aware of lookups,
   so the lookup-engine may resolve properties defined in the
   top level logging configuraton recursively. This applies only
   to the property literal, not the result of lookups used within
   configuration properties.
2. No recursive evaluation of lookup results. While properties are
   a special case, the data returned from other sources must not
   itself be evaluated. This matches the restriction we introduced
   in 2.17.1 for runtime lookups, however it applies everywhere
   and shouldn't result in breaks due to property traversal being
   allowed.

This introduces some level of risk, unlike the changes
in 2.17.1, this update the behavior of all lookups and is not
bounded to less-common runtime lookups.
We do not expect this to break users configurations, however
there are some cases in which we have intentionally changed
behavior:

Given environment `FOO=${lower:BAR}`
Using `<File name="${env:FOO}">` in a configuration will result
in a File appender with the literal name `${lower:BAR}` where
it would have previously evaluated to `bar`. This resolves an
ambiguity problem where those setting environment variables
don't need to be aware of log4j2 or our lookup syntax.