Update PropertiesConfigurationBuilder.java
if we have below properties configured in log4j2.properies files, when we will hit NumberFormatException or there will be no logs for custom level.  This issue is happening because we are trying to convert string "VERBOSE" into integer which is not expected.  It is expected to covert the values which is there in property called intLevel only. 

customLevels = V
customLevel.V.name = VERBOSE
customLevel.V.intLevel = 525