Improves behavior of failing Log4j 1.x builders
If a Builder fails due to a missing property, it return `null`. This
activates the fallback _"instantiate by class name"_ mechanism that will
fail with `NoClassDefFoundException` in the best scenario or will
instantiate a native Log4j 1.x component in the worst scenario.

A better approach, IMHO, is to an empty wrapper, which deactivates the
fallback mechanism, but will be unwrapped to `null` soon after and
removed from the configuration.

This PR also increases the level of status logger messages to `ERROR` if
the condition prevents the instantiation of the component.

Conflicts:
	log4j-1.2-api/src/main/java/org/apache/log4j/builders/BuilderManager.java