[LOG4J-3413] Fixes resolution of non-Log4j properties
For properties that does not start with `log4j` or
`org.apache.logging.log4j` the loosed token-based lookup should not be
applied. There are a couple of exceptions to this rule: the properties
that start with `AsyncLogger` and those related to
`disableThreadContext`.

This PR prevents properties such as `level` to be used as a substitute
for `log4j2.level`

Conflicts:
	log4j-api/src/test/java/org/apache/logging/log4j/util/PropertiesUtilTest.java
	log4j-api/src/test/java/org/apache/logging/log4j/util/PropertySourceTokenizerTest.java