Increase to 50 dependabot proposals
We have 38 upgradeable dependencies. In order to limit the annoyance from Dependabot PRs, I propose to limit the number of `test` scope upgrade proposals to 10 and to give a larger number (50) for the remaining ones.

Another solution would be to prioritize the number of upgrade proposals for `log4j-api` and `log4j-core` and limit those for the remaining modules.