[LOG4J2-2921] Parallel execution for JUnit5 tests
Unit tests in `release-2.x` are executed sequentially in per-test JVMs.
While parallel execution in a single JVM requires a lot of work, this PR
makes a step in this direction: the tests based on JUnit5 in
`log4j-core` (about 50% of the tests) are executed in per-test JVMs, but
in parallel.

This requires only a private/temporary directory for each test to store
its logs (@PrivateDir extension), so that tests using the same file
names don't clash.

As a preliminary step to phase 2 (running multiple JVMs in parallel and
reusing them for multiple tests), this PR also introduces a thread-bound
property source/lookup that can be retrieved with the @TestProperties
annotation (or @SetTestProperty).