Simplify namespace usage in LoggerContextResolver
This uses the ExtensionContext.Namespace API properly by using the same namespace for storing items whether it's for a BeforeAllCallback or BeforeEachCallback. As both of those callbacks have nested ExtensionContexts, this means that getting and putting items in the store work fairly naturally compared to how I initially understood them to work. Now the scope of a LoggerContextSource annotation maps logically to the annotated element itself (the test class or test method).

Signed-off-by: Matt Sicker <mattsicker@apache.org>