Fix ordering of default bundle
At some point, the comparator used for the Ordered annotation changed from reverse to natural. As a bonus, after the Ordered annotation was integrated into the DI system, the ConditionalOnMissingBinding annotation became redundant as that's how they work in the first place unless the binding has a strictly lesser ordinal.

Signed-off-by: Matt Sicker <mattsicker@apache.org>