LOG4J2-2947 - Make LogBuilder methods abstract
The current approach of having no-op default methods is pretty
error-prone because it is easy to forget to override a method when a class
implements the interface. Additionally the documentation stating that the
interface method does nothing is probably quite confusing for users, who
don't care about the implementation details of Log4j 2.