Add scopes to child POM dependencies
This removes the scope of the dependencies from <dependencyManagement>
and adds it to the children POMs.

The reason behind this is that we don't want to force the scope of
transitive dependencies: e.g. `log4j-api` has a `javax.inject`
transitive test dependency. If we set the scope to `provided` in the
parent POM, the test dependency becomes a provided dependency.

This also changes the Jakarta EE 8 API dependencies to Java EE 8 API
dependencies so that version upgrades don't mix up Jakarta EE 8 and
Jakarta EE 9 dependencies (they have the same artifacts).