Adapt tests to Kafka 3.3.1
Even if the `Producer` interface had some breaking changes between
version 1.1.1 and 3.3.1, our implementation extends `KafkaProducer`,
which hides those differences.

The only obstacle for upgrading `kafka-clients` are small differences in
tests. This PR adapts the existing tests to version 3.3.1 and provides a
GitHub action to test the `KafkaAppender` for multiple Kafka versions.