Fix `log4j-api-test` JPMS encapsulation (from master)
In order for `log4j-api-test` to work as JPMS module, the
`ThreadContextUtilityClass` must be moved and `ThreadContext#init()`
must be public. It is already documented as test-only, so this should be
safe.

It also fixes the dependency on `org.junit.platform.commons`, which does
not export its `util` package to non-JUnit modules.