Migrate most of log4j-core to new properties system
This requires the introduction of a few new classes (see ConfigurationResolver and HttpSourceLoader for non-trivial refactoring) for allowing dependency injection or otherwise cleaning up static state. Some APIs are expected to be removed once the migration is completed.

This adds default properties values, though the properties for specifying class names are currently deprecated and expected to be migrated to use the dependency injection system entirely.

This adds a way to specify the logger context name for tests along with updates to ContextSelector to support this.

Signed-off-by: Matt Sicker <mattsicker@apache.org>