Serialize cached location
The current code does not copy the location information if
`includeLocation` is false: the goal was to prevent an expensive call to
`StackLocatorUtil`, but it also erases location obtained through other
means (e.g.  `LogBuilder#withLocation`).

Using `LogBuilder`, however, we can have a situation, where location is
available at no cost.

This PR does not apply to `master`, which does not have serialization
code.