Updates Log4j API doc to reflect Throwable as last parameter feature.
As discussed in #Issue1569 Log4j API doc doesn't mention the feature of passing a Throwable as the last parameter of any of the log methods without adding an extra substituting parameter.

This PR updates Log4j API documentation to reflect the `throwable` as the last parameter feature.
`logger.debug("Logging in user {} with birthday {}", user.getName(), user.getBirthdayCalendar(), exception);`