Increase limit of open Dependabot PRs
This should help avoiding the upper bound deps Maven Enforcer rule,
which requires Dependabot PRs to be merged in an exact order (first a
common transitive dependency, then the main dependency).

By increasing the limit we should have less dead locks that occur when
Dependabot opens a PR for a direct dependency, but can not open a PR for
the transitive dependency.