Deprecate `StatusConsoleListener` filters (#2226)
The configuration attribute `verbose` (i.e., `<Configuration verbose="..."`) and
`StatusConsoleListener` filters are deprecated. This is part of an ongoing effort
to simplify the `StatusLogger`.

`verbose` attribute is deprecated, because it was only effective over
`verboseClasses` and that was hardcoded to a list only containing
`ResourceUtil`. `ResourceUtil` is slightly updated and its `logger.info()`
calls are replaced with `logger.debug()`.