[LOG4J2-3680] Allow deserialization of arrays
The `DefaultObjectInputFilter` is too restrictive and it does not allow
the deserialization of **arrays** of classes in the
`SerializationUtil#REQUIRED_JAVA_CLASSES` and
`SerializationUtil#REQUIRED_JAVA_PACKAGES` white lists.

Most notably `RuntimeException` can not be deserialized, since it
contains an array of `StackTraceElement`s. Both classes are allowed, but
deserialization of the array fails.