Add Logback's throwable-consuming semantics as an option
The semantics used to determine the `Throwable` attached to a log event
changed in Logback 1.1.0 (in the fix to [LOGBACK-873](https://jira.qos.ch/browse/LOGBACK-873)).

When there are not enough arguments to the log call to fill both the
throwable field and the placeholders contained in the message pattern,
Logback 1.1.0 and later prefer to fill the throwable field and leave
some placeholders empty. Log4j Core does the opposite choice: it first
fills all the placeholders and only fills the throwable field if there
is something left.

This change allows `log4j-slf4j-impl` and `log4j-slf4j2-impl` users to
switch between the two behaviors by setting the property
`log4j2.messageFactory` to:

```
org.apache.logging.slf4j.message.ThrowableConsumingMessageFactory
```