Remove CLI tools from `log4j-core` and `log4j-1.2-api`
We remove the command line tools to generate logger wrappers and convert
Log4j 1.x configuration.

The tools have been submitted to Log4j Transform:

 * apache/logging-log4j-transform#92
 * apache/logging-log4j-transform#93

and are not necessary at runtime.

The `Log4j1ConfigurationFactory` that has been removed from
`log4j-1.2-api` is an **older** and deprecated version of the
`PropertiesConfigurationFactory`.