Add support for thread context map propagation
Adds a new `ExtendedThreadContext` SPI class to help external libraries
with an efficient propagation of the `ThreadContextMap`.

Currently integrators can retrieve and set the context map through a
combination of `ThreadContext#getImmutableContext`, `#clear` and
`#putAll`. This inevitable leads to the creation of temporary objects if
the underlying implementation does not use the JDK `Map` class
internally.

As a replacement we provide two `saveMap` and `restoreMap` methods that
can be used to access the `ThreadLocal`s directly. Due to the inherent
unsafety of such operations, these methods are only available from the
`o.a.l.l.spi` package.

We also reimplement `CloseableThreadContext.Instance` to take advantage
of this SPI.