Fixing HTTP redirects based on fragments
The browser does not send the fragment part to the server, when
requesting a page. So all redirects from `a.html#fragment` to
`b.html` will not work.

On the other hand you can redirect `a.html` to `b.html#fragment`,
but you need to add the NE (disables escaping) flag.