Improve plugin descriptor warnings
This improves status logger warnings about:

* missing `Log4j2Plugin.dat` plugin descriptors,
* usage of package scanning.

The warnings are reworded in a way that they are more useful to end-users.
If package scanning is enabled, linkage errors (almost exclusively due to optional dependencies) are logged at a `DEBUG` level instead of `WARN`.

The `PluginManager` tests are revamped to:

- Look for a status logger warning if a deprecated plugin scanning feature is used.
- Look for a status logger warning if no plugin descriptor is available.

This addresses @vlsi suggestions from
https://github.com/apache/jmeter/issues/5937#issuecomment-2271160800 and fixes #2835.

Co-authored-by: Volkan Yazıcı <volkan@yazi.ci>