Avoid surrogate character handling in `JsonWriter`
This change effectively reverts LOG4J2-2998 and #457.

Previously, we were omitting the last character if it is
found to be a high-surrogate. This was to ensure that the
string ending was correctly encoded. Though this doesn't
make sense in the absence of a character encoding. Consider
the following example:

    Is `"\uD800"` a valid JSON document?`

If the character encoding is UTF-8, then it is a valid JSON
document. If it is UTF-16, then it is an *invalid* JSON
document; since `\uD800` is a high-surrogate character and, by
definition, is supposed to be followed by a low-surrogate
character to form a surrogate pair corresponding to a code point.
Hence, given the absence of the character encoding, we should not
attempt to fix surrogate character issues. As a matter of fact,
as of date, neither the reference implementation[1] provides
specialization for surrogates.

See `java.lang.Character` documentation[2] for details.

[1] https://github.com/stleary/JSON-java/blob/14f71274f79ac50afb316845539f4195facc33ee/src/main/java/org/json/JSONTokener.java
[2] https://docs.oracle.com/en%2Fjava%2Fjavase%2F17%2Fdocs%2Fapi%2F%2F/java.base/java/lang/Character.html