Switch MongoDB tests to use Docker
MongoDB is a binary server. The current `log4j-mongodb`
tests download a **generic** binary MongoDB distribution
and try to run it. The distribution is not self-contained
and requires several libraries (e.g., OpenSSL) to be available
on the target host.

Those libraries are not always available in the required version:
e.g., currently MongoDB needs OpenSSL 1, but OpenSSL 3 is bundled
in the most recent Debian.

This PR switches from the binary distribution to the usage of
the **latest** Docker image available.

The advantages of this approach are:

- We always test against the newest server version available.
- The success of the tests does not depend on the libraries
  installed on the host.
- Tests can run in parallel. In the current approach, parallel
  tests failed since each one tried to download MongoDB
  separately.

The main disadvantage is that Docker will be required to test
`log4j-mongodb`. This is the case for the CI, but individual
developers might need to install it too.

Co-authored-by: Volkan Yazıcı <volkan@yazi.ci>