Enforce RTC style on `logging-log4j2` (#3582)
* Enforce RTC style on `logging-log4j2`

Following the [PMC discussion in September 2024](https://lists.apache.org/thread/6gbos0rn3k4y3wjb1hcgnnols4ogqckl), this PR changes the branch protection rules for the `logging-log4j2` repository:

- It requires all changes to `2.x` and `main` to be submitted as PRs.
- It requires all PRs to have at least one positive review to be merged.
- It requires all conversations to be marked as resolved, before a PR is merged.

The current features offered by GitHub do not allow introducing exceptions on the size of the PR or other criteria, so let's try to review 100% of the PRs. Smaller teams manage to do it, so we should be able too.

* Adapt to PMC vote

* Add CodeQL required check