Add `log4j.plugin.enableBndAnnotations` option to `PluginProcessor` (#3258)
* Add `log4j.plugin.enableBndAnnotations` option to `PluginProcessor`

This adds a `log4j.plugin.enableBndAnnotations` option to the `PluginProcessor`. Its default value is inferred from the compiler classpath.

We also rename the `pluginPackage` option to a more coherent `log4j.plugin.package` option.

Closes #3251

* Look for BND annotation on classpath

The processor lives in the annotation processor path, so `Class.forName` does **not** check for the presence of the BND annotation on the classpath.