StatusLogger: Write to stderr by default (#3666)
`StatusLogger` messages used to be written to stderr by default. Since
2.23.0, they are being written to stdout. This is a significant
regression. As a rule, stderr is for humans, and stdout is for programs.
The pipe operator redirects stdout by default, and a CLI utility can
have its output corrupted because Log4j2 emitted a status message to
stdout. For example, consider a program that emits JSON and has its
output piped to `jq`: after upgrading to Log4j 2.23.0 or later, this
program will cease to work correctly unless the StatusLogger is
reconfigured to not emit anything, or is explicitly configured to write
to stderr.

This regression can be fixed by calling the two-parameter constructor
variant of `StatusConsoleListener` from `StatusLogger`. A simple test
case has been added that captures stdout/stderr and ensures that the
default `StatusLogger` is not emitting anything to stdout.

Fixes #3665.