Fixes OSGi descriptor to accept Disruptor 4
Log4j Core added support for Disruptor 4 in `2.23.0` (see #1829), but we forgot to update the OSGi metadata to allow Disruptor 4 in that environment.

This PR expands the accepted Disruptor version range in the OSGi descriptor to ``[3.4,5)``, which partially addresses #3706.

I've also added a small test to try and reproduce the issue from #3706 (where Disruptor version 3 is misidentified as Disruptor 4), and to confirm that async logging works in an OSGi setup. Unfortunately, we couldn't reproduce the bug in our test environment—so that part still needs investigation. However, the version range issue is separate and should now be resolved by this change.